+++ src/main.zig
@@ -13,6 +13,7 @@ comptime {
     _ = exports.turbotoken_encode_utf8_bytes;
     _ = exports.turbotoken_decode_utf8_bytes;
     _ = exports.turbotoken_encode_bpe_from_ranks;
+    _ = exports.turbotoken_count_bpe_from_ranks;
     _ = exports.turbotoken_decode_bpe_from_ranks;
 }